Setup up project to use a cmd folder.
Added embedData internal package since all embed paths are relative to
the package. For instance, moving `main.go` to `cmd/main.go` meant the embedded files
would need to be in `cmd/`

The answer to this was to create the `embedData` package, and import the
files after moving them there.

Changes to be committed:
	modified:   cmd/main.go
	new file:   internal/embedData/embedData.go